sign_up_bot/requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
paramiko==2.10.3
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.1
six==1.16.0
yarl==1.7.2